fix: the channel search flow
1. Fixed the visibility of the plugin to several user roles
2. Display an NPE player label to logged in user
3. Follow the use case on notion for the channel search flow
